Refactor to use remote authentication package
Resolves: DCMAW-7624
Resolves: DCMAW-7724
