--- app/src/androidTest/java/uk/gov/onelogin/home/HomeScreenKtTest.kt
@@ -1,16 +1,0 @@
-package uk.gov.onelogin.home
-
-import dagger.hilt.android.testing.HiltAndroidTest
-import org.junit.Test
-import uk.gov.onelogin.TestCase
-import uk.gov.onelogin.ext.setupComposeTestRule
-
-@HiltAndroidTest
-class HomeScreenKtTest : TestCase() {
-    @Test
-    fun initialisesHomeScreen() {
-        composeTestRule.setupComposeTestRule { _ ->
-            HomeScreen()
-        }
-    }
-}
