--- app/src/androidTest/java/uk/gov/onelogin/home/HomeTestRoutes.kt
@@ -1,9 +1,0 @@
-package uk.gov.onelogin.home
-
-/**
- * Duplicates [HomeRoutes] so that if production code changes, this also requires changing.
- */
-object HomeTestRoutes {
-    const val ROOT: String = "/home"
-    const val START: String = "$ROOT/start"
-}
