+++ app/src/androidTest/java/uk/gov/onelogin/ui/home/HomeScreenKtTest.kt
@@ -1,0 +1,16 @@
+package uk.gov.onelogin.ui.home
+
+import dagger.hilt.android.testing.HiltAndroidTest
+import org.junit.Test
+import uk.gov.onelogin.TestCase
+import uk.gov.onelogin.ext.setupComposeTestRule
+
+@HiltAndroidTest
+class HomeScreenKtTest : TestCase() {
+    @Test
+    fun initialisesHomeScreen() {
+        composeTestRule.setupComposeTestRule { _ ->
+            HomeScreen()
+        }
+    }
+}
