+++ app/src/androidTest/java/uk/gov/onelogin/ui/home/HomeTestRoutes.kt
@@ -1,0 +1,9 @@
+package uk.gov.onelogin.ui.home
+
+/**
+ * Duplicates [HomeRoutes] so that if production code changes, this also requires changing.
+ */
+object HomeTestRoutes {
+    const val ROOT: String = "/home"
+    const val START: String = "$ROOT/start"
+}
