+++ app/src/androidTest/java/uk/gov/onelogin/ui/login/LoginTestRoutes.kt
@@ -1,0 +1,10 @@
+package uk.gov.onelogin.ui.login
+
+/**
+ * Duplicates [LoginRoutes] so that if production code changes, this also requires changing.
+ */
+object LoginTestRoutes {
+    private const val ROOT: String = "/login"
+    const val START: String = "$ROOT/start"
+    const val LOADING: String = "$ROOT/loading"
+}
