--- app/src/main/java/uk/gov/onelogin/components/Color.kt
@@ -1,29 +1,0 @@
-package uk.gov.onelogin.components
-
-import androidx.compose.ui.graphics.Color
-
-val GovGreen = Color(0xFF00703C)
-val IconBackground = Color(0xFF1D70B8)
-val IconForeground = Color.White
-
-val md_theme_light_background = Color(0xFFFBFDF8)
-val md_theme_light_error = Color(0xFFBA1A1A)
-val md_theme_light_onBackground = Color(0xFF000000)
-val md_theme_light_onError = Color(0xFFFFFFFF)
-val md_theme_light_onPrimary = Color(0xFFFFFFFF)
-val md_theme_light_onSecondary = Color(0xFF006D3A)
-val md_theme_light_onSurface = Color(0xFF000000)
-val md_theme_light_primary = Color(0xFF00703C)
-val md_theme_light_secondary = Color(0xFFFFFFFF)
-val md_theme_light_surface = Color(0xFFFBFDF8)
-
-val md_theme_dark_background = Color(0xFF000000)
-val md_theme_dark_error = Color(0xFFFFB4AB)
-val md_theme_dark_onBackground = Color(0xFFFFFFFF)
-val md_theme_dark_onError = Color(0xFF690005)
-val md_theme_dark_onPrimary = Color(0xFFFFFFFF)
-val md_theme_dark_onSecondary = Color(0xFFFFFFFF)
-val md_theme_dark_onSurface = Color(0xFFFFFFFF)
-val md_theme_dark_primary = Color(0xFF008547)
-val md_theme_dark_secondary = Color(0xFF000000)
-val md_theme_dark_surface = Color(0xFF191C19)
