+++ app/src/main/java/uk/gov/onelogin/login/authentication/LoginSessionModule.kt
@@ -1,19 +1,20 @@
 package uk.gov.onelogin.login.authentication
 
+import android.content.Context
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
+import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
-import uk.gov.android.authentication.ILoginSession
+import uk.gov.android.authentication.AppAuthSession
 import uk.gov.android.authentication.LoginSession
-import javax.inject.Inject
-import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
-class LoginSessionModule @Inject constructor() {
-
+object LoginSessionModule {
     @Provides
-    @Singleton
-    fun providesLoginSession(): ILoginSession = LoginSession()
+    fun providesLoginSession(
+        @ApplicationContext
+        context: Context
+    ): LoginSession = AppAuthSession(context)
 }
