+++ app/src/main/res/values-v23/themes.xml
@@ -1,4 +1,4 @@
-<resources xmlns:tools="http://schemas.android.com/tools">
+<resources>
 
     <style name="Theme.OneLogin" parent="Base.Theme.OneLogin">
         <!-- Transparent system bars for edge-to-edge. -->
