+++ app/src/main/res/values/themes.xml
@@ -1,4 +1,4 @@
-<resources xmlns:tools="http://schemas.android.com/tools">
+<resources>
     <!-- Base application theme. -->
     <style name="Base.Theme.OneLogin" parent="Theme.Material3.DayNight.NoActionBar">
         <!-- Customize your light theme here. -->
