+++ app/src/staging/res/values/login_flow.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <string name="openIdConnectBaseUrl" translatable="false">https://oidc.integration.account.gov.uk%1$s</string>
+    <string name="stsUrl" translatable="false">https://token.staging.account.gov.uk</string>
     <string name="openIdConnectClientId" translatable="false">sdJChz1oGajIz0O0tdPdh0CA2zW</string>
     <string name="webBaseUrl" translatable="false">https://mobile.staging.account.gov.uk%1$s</string>
     <string name="webBaseHost" translatable="false">mobile.staging.account.gov.uk</string>
