--- authentication/build.gradle.kts
@@ -1,56 +1,0 @@
-@Suppress("DSL_SCOPE_VIOLATION") // TODO: Remove once KTIJ-19369 is fixed
-plugins {
-    id("com.android.library")
-    id("org.jetbrains.kotlin.android")
-}
-
-android {
-    namespace = "uk.gov.android.authentication"
-    compileSdk = 34
-
-    defaultConfig {
-        minSdk = 29
-
-        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles("consumer-rules.pro")
-    }
-
-    buildTypes {
-        release {
-            isMinifyEnabled = false
-            proguardFiles(
-                getDefaultProguardFile("proguard-android-optimize.txt"),
-                "proguard-rules.pro"
-            )
-        }
-    }
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_1_8
-        targetCompatibility = JavaVersion.VERSION_1_8
-    }
-    kotlinOptions {
-        jvmTarget = "1.8"
-    }
-}
-
-dependencies {
-    listOf(
-        libs.appauth
-    ).forEach(::api)
-
-    listOf(
-        libs.androidx.test.ext.junit,
-        libs.espresso.core
-    ).forEach(::androidTestImplementation)
-
-    listOf(
-        libs.androidx.core.core.ktx,
-        libs.appauth,
-        libs.appcompat,
-        libs.material
-    ).forEach(::implementation)
-
-    listOf(
-        libs.junit
-    ).forEach(::testImplementation)
-}
