+++ features/build.gradle.kts
@@ -1,0 +1,105 @@
+import org.gradle.api.tasks.testing.logging.TestLogEvent
+
+plugins {
+    id("com.android.application")
+    id("org.jetbrains.kotlin.android")
+    id("org.jlleitschuh.gradle.ktlint")
+    id("uk.gov.onelogin.jvm-toolchains")
+}
+
+android {
+    namespace = rootProject.ext["appId"] as String
+    compileSdk = rootProject.ext["compileSdkVersion"] as Int
+
+    defaultConfig {
+        applicationId = rootProject.ext["appId"] as String
+        minSdk = rootProject.ext["minSdkVersion"] as Int
+        targetSdk = rootProject.ext["targetSdkVersion"] as Int
+        versionCode = rootProject.ext["versionCode"] as Int
+        versionName = rootProject.ext["versionName"] as String
+    }
+
+    buildTypes {
+        release {
+            isMinifyEnabled = false
+            proguardFiles(
+                getDefaultProguardFile("proguard-android-optimize.txt"),
+                "proguard-rules.pro"
+            )
+        }
+        debug {
+            enableAndroidTestCoverage = true
+            enableUnitTestCoverage = true
+        }
+    }
+
+    lint {
+        abortOnError = true
+        absolutePaths = true
+        baseline = File("${rootProject.extra["configDir"]}/android/baseline.xml")
+        checkAllWarnings = true
+        checkDependencies = false
+        checkGeneratedSources = false
+        checkReleaseBuilds = true
+        disable.addAll(
+            setOf(
+                "ConvertToWebp",
+                "UnusedIds",
+                "VectorPath"
+            )
+        )
+        explainIssues = true
+        htmlReport = true
+        ignoreTestSources = true
+        ignoreWarnings = false
+        lintConfig = File("${rootProject.extra["configDir"]}/android/lint.xml")
+        noLines = false
+        quiet = false
+        showAll = true
+        textReport = true
+        warningsAsErrors = true
+        xmlReport = true
+    }
+
+    compileOptions {
+        sourceCompatibility = JavaVersion.VERSION_1_8
+        targetCompatibility = JavaVersion.VERSION_1_8
+    }
+    kotlinOptions {
+        jvmTarget = "1.8"
+    }
+
+    testOptions {
+        unitTests.all {
+            it.useJUnitPlatform()
+            it.testLogging {
+                events = setOf(
+                    TestLogEvent.FAILED,
+                    TestLogEvent.PASSED,
+                    TestLogEvent.SKIPPED
+                )
+            }
+        }
+        unitTests {
+            isReturnDefaultValues = true
+            isIncludeAndroidResources = true
+        }
+    }
+}
+
+dependencies {
+
+    listOf(
+        libs.appcompat,
+        libs.androidx.core.core.ktx
+    ).forEach(::implementation)
+
+    listOf(
+        libs.junit
+    ).forEach(::testImplementation)
+
+    listOf(
+        libs.androidx.test.ext.junit,
+        libs.espresso.core
+    ).forEach(::androidTestImplementation)
+}
