+++ gradle/libs.versions.toml
@@ -92,6 +92,7 @@ uiautomator = { group = "androidx.test.uiautomator", name = "uiautomator", versi
 ##                                                                              ⬆ = "2.3.0-alpha03" }
 ##                                                                              ⬆ = "2.3.0-alpha04" }
 ##                                                                              ⬆ = "2.3.0-alpha05" }
+authentication = "uk.gov.android:authentication:0.4.1"
 viewbinding = { group = "androidx.databinding", name = "viewbinding", version.ref = "androidx-databinding" }
 appauth = { group = "net.openid", name = "appauth", version = "0.11.1" }
 androidx-core-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "core-ktx" }
